requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
python-dotenv==0.15.0
django-cors-headers==3.5.0
